Add UDFs to glossary
If terms as common as "API" are included in the glossary, the not so common abbreviation "UDF" should be explained there too.
In general the concept of UDFs is worthy of a brief description there IMO - I tried to come up with something, please replace or amend as you see fit.