Harmonise definition and example of openeo.debug topic
The definition and example for the `openeo.debug` topic were completely out of sync concerning where to gather the debug message, process information etc.: collected in a common `debug` object or as direct children of the payload. I like the latter better, so I made the definition match the example.

With one exception: The `job_id` was specified as "required" in the definition but left out in the example. Here I adhered to the existing definition and added a `job_id` property to the example JSON because it was most likely just forgotten.